Create README.md
Hi everyone, I'm Jonas and i wanted to make a detailed freeware 787 for x-plane. i want this to be a nice study level 787-9 that people can enjoy flying with.
My goal is to finish it as soon as possible, i have a few developers working on it.
If you're a developer, coder  animator dm me on discord= Brussels_aviation.